Add support for non-JUint tests. Refactor TestSuite (create TestHelpers). Add FunTestSuite for ScalaTest FunSuites.
